+++ client_src/js/system.runner.js
@@ -4,7 +4,9 @@ import ClientSettings from "./system/client-settings";
 import {
   localStorageKeys,
   serverActionsElements,
-  serverDatabaseKeys
+  serverDatabaseKeys,
+  userActionElements,
+  returnSaveBtn
 } from "./system/server.options";
 import {
   generateLogDumpFile,
@@ -17,7 +19,12 @@ import {
   checkForOctoFarmUpdates,
   grabOctoFarmLogList,
   startUpdatePageRunner,
-  renderSystemCharts
+  renderSystemCharts,
+  createNewUser,
+  editUser,
+  deleteUser,
+  resetUserPassword,
+  fillInEditInformation
 } from "./system/server.actions";
 import { serverBootBoxOptions } from "./system/utils/bootbox.options";
 
@@ -90,4 +97,71 @@ serverActionsElements.CHECK_OCTOFARM_UPDATES.addEventListener("click", async (e)
 serverActionsElements.SAVE_CLIENT_SETTINGS.addEventListener("click", async (e) => {
   // Validate Printer Form, then Add
   await ClientSettings.update();
+});
+
+// Setup user create btn
+userActionElements.createUserBtn.addEventListener("click", (e) => {
+  userActionElements.createUserFooter.innerHTML = `
+    ${returnSaveBtn()}
+  `;
+  const userActionSave = document.getElementById("userActionSave");
+  userActionSave.addEventListener("click", async () => {
+    await createNewUser();
+  });
+});
+
+// Setup user password resets
+const passResetButtons = document.querySelectorAll('*[id^="resetPasswordBtn-"]');
+passResetButtons.forEach((btn) => {
+  const split = btn.id.split("-");
+  btn.addEventListener("click", (e) => {
+    userActionElements.resetPasswordFooter.innerHTML = `
+    ${returnSaveBtn()}
+  `;
+    const userActionSave = document.getElementById("userActionSave");
+    userActionSave.addEventListener("click", async () => {
+      await resetUserPassword(split[1]);
+    });
+  });
+});
+
+// Setup user edits
+const userEditButtons = document.querySelectorAll('*[id^="editUserBtn-"]');
+userEditButtons.forEach((btn) => {
+  const split = btn.id.split("-");
+  btn.addEventListener("click", async (e) => {
+    userActionElements.editUserFooter.innerHTML = `
+    ${returnSaveBtn()}
+  `;
+    await fillInEditInformation(split[1]);
+    const userActionSave = document.getElementById("userActionSave");
+    userActionSave.addEventListener("click", async () => {
+      await editUser(split[1]);
+    });
+  });
+});
+
+// Setup user deletes
+const deleteButtons = document.querySelectorAll('*[id^="deleteUserBtn-"]');
+deleteButtons.forEach((btn) => {
+  const split = btn.id.split("-");
+  btn.addEventListener("click", (e) => {
+    deleteUser(split[1]);
+  });
+});
+
+$("#userCreateModal").on("hidden.bs.modal", function (e) {
+  const userActionSave = document.getElementById("userActionSave");
+  userActionElements.userCreateMessage.innerHTML = "";
+  userActionSave.remove();
+});
+$("#usersResetPasswordModal").on("hidden.bs.modal", function (e) {
+  const userActionSave = document.getElementById("userActionSave");
+  userActionElements.userResetMessage.innerHTML = "";
+  userActionSave.remove();
+});
+$("#userEditModal").on("hidden.bs.modal", function (e) {
+  const userActionSave = document.getElementById("userActionSave");
+  userActionElements.userEditMessage.innerHTML = "";
+  userActionSave.remove();
 });