+++ RELEASENOTES.md
@@ -28,6 +28,8 @@
  - New server side settings to enable and diable any monitoring views... You can now turn on/off each view in System -> Server -> Monitoring Views.
  - Added #299: New View combined mode, is a bulked out list view. Pretty much a combination of Panel/List/Camera views rolled into one super view.
  - New section in System for Release Notes. This will display the current version changes from the previous version.
+ - New user CRUD endpoints: /users/users. Required administrator rights to access.
+ - New system tab called "Users". Can created, edit, read, update all users. There is also a password reset option.
 
 ### Changed